refactor: 패키지 명 변경 및 movable 함수 추가
