--- src/main/java/chess/Role.java
@@ -1,10 +1,0 @@
-package chess;
-
-public class Role {
-
-    private String role;
-
-    public Role(String role) {
-        this.role = role;
-    }
-}
