--- src/main/java/chess/model/Rank.java
@@ -1,44 +1,0 @@
-package chess.model;
-
-import java.util.Arrays;
-import java.util.List;
-
-public enum Rank {
-
-    ONE(1),
-    TWO(2),
-    THREE(3),
-    FOUR(4),
-    FIVE(5),
-    SIX(6),
-    SEVEN(7),
-    EIGHT(8)
-    ;
-
-    private static final String INVALID_RANK_LOCATION_FORMAT = "잘못된 위치입니다. %d";
-
-    private final int col;
-
-    Rank(int col) {
-        this.col = col;
-    }
-
-    public static Rank of(int col) {
-        return Arrays.stream(values())
-                .filter(rank -> rank.col == col)
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException(String.format(INVALID_RANK_LOCATION_FORMAT, col)));
-    }
-
-    public Rank add(int col) {
-        return Rank.of(this.col + col);
-    }
-
-    public int calculateGap(Rank target) {
-        return target.col - this.col;
-    }
-
-    public static List<Rank> emptyBaseLine() {
-        return List.of(SIX, FIVE, FOUR, THREE);
-    }
-}
