+++ src/main/java/chess/piece/King.java
@@ -1,4 +1,6 @@
-package chessrefactor.piece;
+package chess.piece;
+
+import chess.square.Square;
 
 public class King extends Piece {
 
@@ -9,5 +11,10 @@ public class King extends Piece {
     @Override
     public String name() {
         return "k";
+    }
+
+    @Override
+    public boolean movable(Square source, Square target) {
+        return false;
     }
 }
