+++ src/main/java/chess/piece/Pawn.java
@@ -1,4 +1,6 @@
-package chessrefactor.piece;
+package chess.piece;
+
+import chess.square.Square;
 
 public class Pawn extends Piece {
 
@@ -9,5 +11,10 @@ public class Pawn extends Piece {
     @Override
     public String name() {
         return "p";
+    }
+
+    @Override
+    public boolean movable(Square source, Square target) {
+        return false;
     }
 }
