+++ src/main/java/chess/piece/Rook.java
@@ -1,4 +1,6 @@
-package chessrefactor.piece;
+package chess.piece;
+
+import chess.square.Square;
 
 public class Rook extends Piece {
 
@@ -9,5 +11,10 @@ public class Rook extends Piece {
     @Override
     public String name() {
         return "r";
+    }
+
+    @Override
+    public boolean movable(Square source, Square target) {
+        return false;
     }
 }
