--- src/main/java/chess/state/GameState.java
@@ -1,16 +1,0 @@
-package chess.state;
-
-import chess.controller.ChessService;
-import chess.model.GameCommand;
-import java.util.List;
-
-public interface GameState {
-
-    GameState changeStatus(GameCommand command);
-
-    boolean isEnd();
-
-    GameState execute(ChessService service, List<String> sqaures);
-
-    boolean isStatus();
-}
