--- src/main/java/chess/state/Running.java
@@ -1,20 +1,0 @@
-package chess.state;
-
-import chess.model.GameCommand;
-
-public abstract class Running implements GameState {
-
-    @Override
-    public GameState changeStatus(GameCommand command) {
-        if (command.isStart()) {
-            return new Start();
-        }
-        if (command.isMove()) {
-            return new Move();
-        }
-        if (command.isStatus()) {
-            return new Status();
-        }
-        return new End();
-    }
-}
