--- src/main/java/chessrefactor/piece/Bishop.java
@@ -1,13 +1,0 @@
-package chessrefactor.piece;
-
-public class Bishop extends Piece {
-
-    public Bishop(Color color) {
-        super(color);
-    }
-
-    @Override
-    public String name() {
-        return "b";
-    }
-}
