--- src/main/java/chessrefactor/piece/Empty.java
@@ -1,13 +1,0 @@
-package chessrefactor.piece;
-
-public class Empty extends Piece {
-
-    public Empty(Color color) {
-        super(color);
-    }
-
-    @Override
-    public String name() {
-        return ".";
-    }
-}
